+++ module-apps/application-meditation/widgets/TimerProperty.hpp
@@ -12,19 +12,21 @@
 
 namespace gui
 {
-    class TimerSetter : public Rect
+    class TimerProperty : public Rect
     {
         class State
         {
-            static constexpr int minimalValue          = 10;
-            static constexpr int maximalValue          = 90;
-            static constexpr int defaultIncrementValue = 15;
+            static constexpr int counterMaxDigits             = 2;
+            static constexpr int minimalValue                 = 1;
+            static constexpr int maximalValue                 = 99;
+            static constexpr int defaultMeditationTime        = 15;
+            static constexpr std::array<const int, 4> timeArr = {15, 30, 60, 90};
 
             bool resetValueOnNumeric = true;
-            int timeInMinutes        = defaultIncrementValue;
+            int timeInMinutes        = defaultMeditationTime;
 
           public:
-            std::chrono::minutes getTime() const noexcept
+            [[nodiscard]] std::chrono::minutes getTime() const noexcept
             {
                 return std::chrono::minutes{timeInMinutes};
             }
@@ -44,13 +46,14 @@ namespace gui
         Label *timeUnitLabel = nullptr;
 
         void build();
+        void setMeditationTime();
 
       public:
-        TimerSetter(Item *parent, const uint32_t x, const uint32_t y, const uint32_t w, const uint32_t h);
+        TimerProperty(Item *parent, const uint32_t x, const uint32_t y, const uint32_t w, const uint32_t h);
 
         bool onFocus(bool isFocused) final;
         bool onInput(const InputEvent &inputEvent) final;
-        [[nodiscard]] std::chrono::seconds getTime() noexcept;
+        [[nodiscard]] std::chrono::minutes getTime() noexcept;
     };
 
 } // namespace gui