+++ module-bsp/bsp/lpm/bsp_lpm.hpp
@@ -18,6 +18,16 @@ namespace bsp {
             Suspend
 
         };
+        enum class CpuFrequency
+        {
+            Level_1, // 12 MHz
+            Level_2, // 24 MHz
+            Level_3, // 66 MHz
+            Level_4, // 132 MHz
+            Level_5, // 264 MHz
+            Level_6  // 528 MHz
+        };
+
         LowPowerMode()          = default;
         virtual ~LowPowerMode() = default;
 
@@ -28,7 +38,7 @@ namespace bsp {
 
         virtual int32_t PowerOff() = 0;
         virtual int32_t Reboot() = 0;
-
+        virtual void SetCpuFrequency(CpuFrequency freq) = 0;
 
     protected:
       Mode currentMode = Mode::FullSpeed;